flask-backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy<2.0
graphene==3.3
graphene-sqlalchemy==3.0.0b4
graphql-relay==3.2.0
promise==2.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.0
PyJWT==2.10.1
requests==2.31.0